Add resume option.
Now that we've got sane arg parsing and the ability to
restore from an SVG file; offer this to users through
a command line argument.

Can only accept a boolean resume token for now. Ideally
the user would be able to specify a file to resume from,
rather than just assuming the default save file.